dart/docs/requirements.txt
# Requirements for building documentation
# Install with: pip install -r requirements.txt

sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0
myst-parser>=2.0.0
PyYAML>=6.0
